# feat: optimize calculateSampleRates when there's one bucket