@kentquirk 2024-07-15
I'd be interested to know how you know "it's not uncommon" to have a single entry -- I don't think I've ever seen it in any of the sample files I've worked with. Far more common is way too many keys.

I don't object to this optimization (other than the generic "it adds complexity", but it's really isolated here), but I'm not sure if it's actually going to be seen in practice.