@cartermp 2024-07-15
@kentquirk It's actually a way that we onboard customers now, the final dynamic sampling rule we give people is the one I mention in the PR title.

> Far more common is way too many keys.

I did think about this one too and couldn't come up with anything that was meaningfully better. I would also guess that when there's a lot of key fields listed, maybe the most time is spent in creating the key rather than rebalancing.